Refactor SpotlightMap
